Fix stupid export bug

+++ src/exports/tasks.py
@@ -15,15 +15,18 @@ def process_export(export_id):
 
     # Cleanup oldest export when there are too many
     owner = export.owner
-    oldest_export_index = settings.EXPORTS_TO_KEEP - 1
-    oldest_export = Export.objects \
+    oldest_exports_qs = Export.objects \
         .filter(owner=owner) \
         .order_by('-created_on') \
-        .all()[oldest_export_index]
-    Export.objects \
-        .filter(owner=owner) \
-        .filter(created_on__lt=oldest_export.created_on) \
-        .delete()
+        .all()
+    oldest_exports = list(oldest_exports_qs)
+    if len(oldest_exports) > settings.EXPORTS_TO_KEEP:
+        oldest_export_index = settings.EXPORTS_TO_KEEP - 1
+        oldest_export = oldest_exports[oldest_export_index]
+        Export.objects \
+            .filter(owner=owner) \
+            .filter(created_on__lt=oldest_export.created_on) \
+            .delete()
 
     export.status = 'processing'
     export.save()
